"realistic friction" explanation is wrong.
The drag equation is F = 1/2 ρ v² C A where
- ρ is the fluid density
- v is the relative velocity to the fluid
- C is a drag coefficient
- A the cross section area

The mass of the particle does not appear here.

However F = ma, thus a=F/m, an so v+=DELTA*F/m which could make the mass appear.
But the explanation is still wrong : we have a mass of 1/0.05=20 instead of 0.05... And we would need to reformulate again to move this factor outside of the force, but clearly this was not taken in account before on other force integration. The mass is assumed to be 1.

Therefore 0.05 is just an arbitrary factor.